requirements.txt
Flask
google-generativeai
firebase-admin
python-dotenv